Add files via upload
Added some customized message and also added additional information on when it's going to expire. This script may be useful in scenarios like users with account expire attribute is used, so that contractors or third party vendor can connect with their customer's respective manager to update extend their account as needed.